+++ app/src/main/res/navigation/ranking.xml
@@ -15,6 +15,10 @@
         <action
             android:id="@+id/action_ranking_to_drink_detail"
             app:destination="@id/drink_detail_fragment" />
+
+        <action
+            android:id="@+id/action_ranking_to_swipe_sorting_dialog"
+            app:destination="@+id/swipe_sorting_dialog_fragment" />
     </fragment>
 
     <fragment
@@ -26,4 +30,8 @@
             android:name="drinkDetail"
             app:argType="jp.kuaddo.tsuidezake.model.DrinkDetail" />
     </fragment>
+
+    <dialog
+        android:id="@+id/swipe_sorting_dialog_fragment"
+        android:name="jp.kuaddo.tsuidezake.ui.ranking.SwipeSortingDialogFragment" />
 </navigation>
