+++ app/src/main/res/values/styles.xml
@@ -1,7 +1,13 @@
 <resources>
+
     <style name="AppTheme" parent="Theme.MaterialComponents.Light.DarkActionBar">
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
+    </style>
+
+    <style name="AppTheme.FullScreenDialog">
+        <item name="android:windowIsFloating">false</item>
+        <item name="android:windowBackground">@color/loading_background</item>
     </style>
 </resources>
